requirements.txt
transformers
torch
setuptools
pandas
python-dotenv
langchain
dspy-ai
openai
jupyter 
ipywidgets